feat: complete Day-37
refers to: https://www.hackingwithswift.com/100/swiftui/37
