+++ SwiftUI100Days/Project/iExpense/AddView.swift
@@ -8,33 +8,35 @@
 import SwiftUI
 
 struct AddView: View {
+    var expenses: Expenses
+    @Environment (\.dismiss) var dismiss
+
     @State private var name = ""
     @State private var type = "Personal"
     @State private var amount = 0.0
 
-    @State private var types = ["Business", "Personal"]
-    var expenses: Expenses
+    let types = ["Business", "Personal"]
 
     var body: some View {
         NavigationStack {
             Form {
                 TextField("Name", text: $name)
-                
+
                 Picker("Type", selection: $type) {
                     ForEach(types, id: \.self) {
                         Text($0)
                     }
                 }
-                
+
                 TextField("Amount", value: $amount, format: .currency(code: "USD"))
                     .keyboardType(.decimalPad)
             }
             .navigationTitle("Add new expense")
-            .navigationBarTitleDisplayMode(.inline)
             .toolbar {
                 Button("Save") {
                     let item = ExpenseItem(name: name, type: type, amount: amount)
                     expenses.items.append(item)
+                    dismiss()
                 }
             }
         }
